Se generaron tablas plantillas y se crean segun plantilla deseada

+++ crud_tipo_usuarios.php
@@ -1,22 +1,23 @@
 <?php
 session_start();
-$arreglo = json_decode(file_get_contents('php://input'), true);
-include("conexionGhoner.php");
-header('Content-Type: application/json');
-$accion = $arreglo['accion'];
-$nuevo = $arreglo['nuevo_tipo'];
+if(isset($_SESSION['nombre'])){
+        $arreglo = json_decode(file_get_contents('php://input'), true);
+        include("conexionGhoner.php");
+        header('Content-Type: application/json');
+        $accion = $arreglo['accion'];
+        $nuevo = $arreglo['nuevo_tipo'];
 
-$resultado = "";  
+        $resultado = "";  
 
-switch ($accion) {
-    case 'insertar':
+        switch ($accion) {
+        case 'insertar':
         $insertar = "INSERT INTO tipo_usuario (tipo_usuario) VALUES ('$nuevo')";
         $query=$conexion->query($insertar);
         if($query){
             $resultado = $query;
         }
         break;
-    case 'eliminar':
+        case 'eliminar':
             $usuario = $arreglo['usuario'];
             $delete = "DELETE FROM tipo_usuario WHERE tipo_usuario='$usuario'";
             $query=$conexion->query($delete);
@@ -24,12 +25,15 @@ switch ($accion) {
                 $resultado = $query;
             }
             break;
-    
-    default:
+
+        default:
         # code...
         break;
-}
+        }
 
-     
-echo json_encode($resultado);
+
+        echo json_encode($resultado);
+}else{
+    header("Location:index.php");
+}
 ?>
